Refactor face detection options, incorporate DeepFace updates
